Resume_website/requirements.txt
tkinter==8.6
streamlit==1.12.2
pillow==9.2.0